requirements.txt
alabaster==1.0.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.1
docutils==0.21.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
libsass==0.23.0
MarkupSafe==3.0.2
packaging==24.2
Pygments==2.18.0
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx_theme_pd==0.8.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-sass==0.3.4
sphinxcontrib-serializinghtml==2.0.0
urllib3==2.3.0
